+++ e2e-tests/handcraft/petstore/show_pet_by_id.test.ts
@@ -1,5 +1,5 @@
-import { client } from "../client.ts";
-import { assertEquals } from "../deps.ts";
+import { client } from "../../client.ts";
+import { assertEquals } from "../../deps.ts";
 
 const endpoint = "pets";
